requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
slowapi>=0.1.9

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv

# ── OpenAI + LangChain ────────────────────────────────────────────────────────
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-core>=0.2.0

# ── Vector store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Retrieval ─────────────────────────────────────────────────────────────────
rank-bm25>=0.2.2
sentence-transformers>=2.7.0

# ── Cache ─────────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── Ingestion (unstructured.io) ───────────────────────────────────────────────
unstructured[all-docs]>=0.14.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity>=8.2.0
numpy>=1.26.0
